+++ pkg/logged-cluster/interface.go
@@ -5,7 +5,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-const LoggingEnabledDefault = true
+// LoggingEnabledDefault defines if WCs logs are collected by default
+const LoggingEnabledDefault = false
 
 // Interface contains the definition of functions that can differ between each type of cluster
 type Interface interface {
@@ -15,8 +16,6 @@ type Interface interface {
 	GetEnableLoggingFlag() bool
 	GetClusterName() string
 	GetInstallationName() string
-	GetRegion() string
-	GetCloudDomain() string
 	AppConfigName(app string) string
 	ObservabilityBundleConfigLabelName(config string) string
 	GetObject() client.Object